Move login-form state to Login
