+++ src/containers/login.js
@@ -4,15 +4,68 @@ import { connect } from 'react-redux'
 
 import LoginForm from '../components/login-form'
 
+import { isValidEmail } from '../lib'
+
 class Login extends React.Component {
   constructor(props) {
     super(props)
+    this.state = {
+      presubmitErrors: null
+    }
     this.handleSubmit = this.handleSubmit.bind(this)
+    this.errorList = this.errorList.bind(this)
   }
 
-  handleSubmit(fields) {
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.errors) {
+      this.setState({ presubmitErrors: null })
+    }
+    this.password.value = ''
+  }
+
+  /**
+   * Validates the form for client side errors.
+   * If valid returns true otherwise false.
+   * If errors it will update the local state `presubmitErrors`
+   * @returns {boolean}
+   */
+  validateForm() {
+    const { email, password } = this
+    const errors = []
+    if (!isValidEmail(email.value)) {
+      if (!this.email.value.trim().length) {
+        errors.push({ message: 'Missing required entry for the Email field.' })
+      } else {
+        errors.push({ message: 'Invalid entry for the Email field.' })
+      }
+    }
+    if (!password.value.trim().length) {
+      errors.push({ message: 'Missing required entry for the Password field.' })
+    }
+    if (errors.length) {
+      this.setState({ presubmitErrors: errors })
+    }
+    return !errors.length
+  }
+
+  /**
+   * Get the current errors as a jsx array.
+   * @returns {Array} an jsx array of errors
+   */
+  errorList() {
+    const errors = this.state.presubmitErrors || this.props.loginErrors || []
+    return errors.map((error, idx) => <li key={idx}>{error.message}</li>)
+  }
+
+  handleSubmit(event) {
+    event.preventDefault()
+    if (!this.validateForm()) return
     // Not implemented yet
-    alert(fields)
+    const fields = {
+      email: this.email.value,
+      password: this.password.value
+    }
+    console.log(fields)
     // this.props.dispatch(sessionActions.login(fields))
   }
 
@@ -20,9 +73,9 @@ class Login extends React.Component {
     return (
       <div className='moveon-petitions'>
         <LoginForm
-          errors={this.props.loginErrors}
-          onSubmit={this.handleSubmit}
-          isSubmitting={this.props.isSubmitting}
+          errorList={this.errorList}
+          handleSubmit={this.handleSubmit}
+          setRef={input => input && (this[input.name] = input)}
         />
       </div>
     )
