+++ test/pages/login.js
@@ -8,6 +8,7 @@ import { createMockStore } from 'redux-test-utils'
 import Login from '../../src/containers/login'
 import LoginForm from '../../src/components/login-form'
 
+const empty = createMockStore({ userStore: {} })
 
 describe('<Login />', () => {
   const baseStore = createMockStore({ userStore: {} })
@@ -19,5 +20,21 @@ describe('<Login />', () => {
 
   it.skip('can log the user in (TODO)', () => {
     // Implement when logging in is implemented
+  })
+
+  it('displays errors when required fields are missing', () => {
+    const login = mount(<Provider store={empty} children={<Login />} />)
+    login.find('form').simulate('submit')
+    expect(login.find('.errors').find('li').length).to.equal(2)
+  })
+
+  it('validates email', () => {
+    const login = mount(<Provider store={empty} children={<Login />} />)
+    login.find('input[name="email"]').node.value = 'foo'
+    login.find('input[name="password"]').node.value = 'bar'
+    login.find('form').simulate('submit')
+    expect(login.find('.errors').text()).to.equal(
+      'Invalid entry for the Email field.'
+    )
   })
 })
